requirements.txt
# Core Django & Deployment
Django>=5.0.0
gunicorn
whitenoise>=6.0.0

# Database
psycopg2-binary
dj-database-url

# Rest Framework & Auth
djangorestframework
djangorestframework-simplejwt
django-cors-headers
bcrypt>=4.2.0
PyJWT>=2.10.0
cryptography
django-ratelimit

# Utilities
python-dotenv
requests>=2.32.0
Pillow>=10.4.0
pyotp>=2.9.0
